+++ packages/components/pagination/src/index.ts
@@ -1,0 +1,3 @@
+export { Pagination } from './Pagination';
+export type { PaginationProps } from './Pagination';
+export { getExpectedItemsOfActivePage } from './utils';