+++ packages/components/menu/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.3](https://github.com/contentful/forma-36/compare/@contentful/f36-menu@4.2.2...@contentful/f36-menu@4.2.3) (2022-02-17)
+
+**Note:** Version bump only for package @contentful/f36-menu
+
+
+
+
+
+## [4.2.2](https://github.com/contentful/forma-36/compare/@contentful/f36-menu@4.2.1...@contentful/f36-menu@4.2.2) (2022-02-09)
+
+
+### Bug Fixes
+
+* menu item focus styles ([#1871](https://github.com/contentful/forma-36/issues/1871)) ([e56531e](https://github.com/contentful/forma-36/commit/e56531e39cc87ccc3ee12d2978ec67a0b30ab25c))
+
+
+
+
+
+## [4.2.1](https://github.com/contentful/forma-36/compare/@contentful/f36-menu@4.2.0...@contentful/f36-menu@4.2.1) (2022-02-09)
+
+**Note:** Version bump only for package @contentful/f36-menu
+
+
+
+
+
+# [4.2.0](https://github.com/contentful/forma-36/compare/@contentful/f36-menu@4.1.3...@contentful/f36-menu@4.2.0) (2022-02-09)
+
+
+### Features
+
+* by default render Popover only when it's open ([#1873](https://github.com/contentful/forma-36/issues/1873)) ([48a511b](https://github.com/contentful/forma-36/commit/48a511bc48c17d3e4bf0cbfb8d16da8c3cbc29b2))
+
+
+
+
+
+## [4.1.3](https://github.com/contentful/forma-36/compare/@contentful/f36-menu@4.1.2...@contentful/f36-menu@4.1.3) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-menu
+
+
+
+
+
 ## [4.1.2](https://github.com/contentful/forma-36/compare/@contentful/f36-menu@4.1.1...@contentful/f36-menu@4.1.2) (2022-01-31)
 
 **Note:** Version bump only for package @contentful/f36-menu